Fix the features
